+++ packages/malloy/src/connection/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './base_connection';
