+++ packages/malloy/src/connection/types.ts
@@ -1,0 +1,156 @@
+import {RunSQLOptions} from '../run_sql_options';
+import {
+  Annotation,
+  MalloyQueryData,
+  QueryDataRow,
+  QueryRunStats,
+  SQLBlock,
+  StructDef,
+} from '../model/malloy_types';
+import {Dialect} from '../dialect';
+
+/**
+ * Options passed to fetchSchema methods.
+ */
+export interface FetchSchemaOptions {
+  // Fetch a fresh copy of the schema instead of using cache
+  refreshTimestamp?: number;
+  /* This is an experimental feature */
+  modelAnnotation?: Annotation;
+}
+
+/**
+ * An object capable of reading schemas for given table names.
+ */
+export interface InfoConnection {
+  // TODO should we really be exposing StructDef like this?
+  // TODO should this be a Map instead of a Record in the public interface?
+  /**
+   * Fetch schemas for multiple tables.
+   *
+   * @param tables The names of tables to fetch schemas for, represented
+   * as a map of keys to table names.
+   * @return A mapping of table keys to schemas.
+   */
+  fetchSchemaForTables(
+    tables: Record<string, string>,
+    options: FetchSchemaOptions
+  ): Promise<{
+    schemas: Record<string, StructDef>;
+    errors: Record<string, string>;
+  }>;
+
+  /**
+   * Fetch schemas an SQL blocks
+   *
+   * @param block The SQL blocks to fetch schemas for.
+   * @return A mapping of SQL block names to schemas.
+   */
+
+  fetchSchemaForSQLBlock(
+    block: SQLBlock,
+    options: FetchSchemaOptions
+  ): Promise<
+    | {structDef: StructDef; error?: undefined}
+    | {error: string; structDef?: undefined}
+  >;
+
+  /**
+   * The name of the connection.
+   */
+  get name(): string;
+}
+
+export type ConnectionParameterValue =
+  | string
+  | number
+  | boolean
+  | Array<ConnectionParameterValue>;
+
+export interface ConnectionParameter {
+  name: string;
+  label: string;
+  type: 'string' | 'number' | 'boolean';
+  isOptional?: boolean;
+  isSecret?: boolean;
+  defaultValue?: ConnectionParameterValue;
+}
+
+export type ConnectionConfigSchema = ConnectionParameter[];
+
+export interface ConnectionConfig {
+  name: string;
+  [key: string]: ConnectionParameterValue | undefined;
+}
+
+export interface ConnectionFactory {
+  connectionName: string;
+  configSchema: ConnectionConfigSchema;
+  createConnection(
+    connectionConfig: ConnectionConfig,
+    dialectRegistrar?: (dialect: Dialect) => void
+  ): Connection & TestableConnection;
+}
+
+/**
+ * An object capable of running SQL.
+ */
+export interface Connection extends InfoConnection {
+  /**
+   * Run some SQL and yield results.
+   *
+   * @param sql The SQL to run.
+   * @param options.pageSize Maximum number of results to return at once.
+   * @return The rows of data resulting from running the given SQL query
+   * and the total number of rows available.
+   */
+  runSQL(sql: string, options?: RunSQLOptions): Promise<MalloyQueryData>;
+
+  // TODO feature-sql-block Comment
+  isPool(): this is PooledConnection;
+
+  canPersist(): this is PersistSQLResults;
+
+  canStream(): this is StreamingConnection;
+
+  close(): Promise<void>;
+
+  estimateQueryCost(sqlCommand: string): Promise<QueryRunStats>;
+
+  get dialectName(): string;
+}
+
+// TODO feature-sql-block Comment
+export interface TestableConnection extends Connection {
+  // TODO feature-sql-block Comment
+  test(): Promise<void>;
+}
+
+export interface PooledConnection extends Connection {
+  // Most pool implementations require a specific call to release connection handles. If a Connection is a
+  // PooledConnection, drain() should be called when connection usage is over
+  drain(): Promise<void>;
+  isPool(): true;
+}
+
+export interface PersistSQLResults extends Connection {
+  manifestTemporaryTable(sqlCommand: string): Promise<string>;
+}
+
+export interface StreamingConnection extends Connection {
+  runSQLStream(
+    sqlCommand: string,
+    options?: RunSQLOptions
+  ): AsyncIterableIterator<QueryDataRow>;
+}
+
+/**
+ * A mapping of connection names to connections.
+ */
+export interface LookupConnection<T extends InfoConnection> {
+  /**
+   * @param connectionName The name of the connection for which a `Connection` is required.
+   * @return A promise to a `Connection` for the connection named `connectionName`.
+   */
+  lookupConnection(connectionName?: string): Promise<T>;
+}
